requirements.txt
celery==5.2.7
redis==4.3.4
pydantic==1.9.2
loguru==0.6.0
omegaconf==2.2.3
open-clip-torch==1.3.0
einops==0.4.1
huggingface-hub==0.9.0
pytorch-lightning ==1.7.2
transformers==4.21.1